refactor: move makeTimeout to home utils
