+++ client/src/feat/home/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./export-data";
+export * from "./make-timeout";
 export * from "./read-json";
